The nested flexbox bug rears its head again. This should hopefully fix the issues. Essentially, a problem arises when doing dimension calculations. As I bubble down and apply flexbox values, I'm changing each flexbox width/height, etc.
The end result therefore might be a flexbox layout where an inordinate amount of whitespace exists. This can't simply be fixed by bubbling up the DOM, because the same issue would still remain, only backwards.

The fix is to (unfortunately), rubber-band both down the DOM, and back up. This is expensive, but necessary for flexbox layouts.

So I'm introducing a bubbleUp method, which checks each parent for existence of an element with a FLX_DOM_ID expando. If found, it re-applies its flex properties.